entrycheck: Raise for Styx 5.0

+++ serendipity_event_entrycheck/serendipity_event_entrycheck.php
@@ -20,15 +20,15 @@ class serendipity_event_entrycheck extends serendipity_event
         $propbag->add('author',        'Garvin Hicking, Gregor Voeltz, Ian Styx');
         $propbag->add('version',       '1.21');
         $propbag->add('requirements',  array(
-            'serendipity' => '1.6',
-            'smarty'      => '2.6.7',
-            'php'         => '5.1.0'
+            'serendipity' => '5.0',
+            'smarty'      => '4.1',
+            'php'         => '8.2'
         ));
         $propbag->add('event_hooks',    array(
             'backend_entry_updertEntry' => true,
             'backend_entry_checkSave'   => true,
             'backend_entryform'         => true,
-            'css_backend'               => true
+            'css'                       => true
         ));
         $propbag->add('groups', array('BACKEND_EDITOR'));
         $propbag->add('configuration', array('emptyCategories', 'emptyTitle', 'emptyBody', 'emptyExtended', 'defaultCat', 'locking'));
@@ -120,8 +120,9 @@ class serendipity_event_entrycheck extends serendipity_event
             // Entry is locked
 
             // Check if it should timeout after one hour
-            if ($locked['locked'] < (time() - 3600) || $serendipity['GET']['unlock'] == 'true') {
+            if ($locked['locked'] < (time() - 3600) || (isset($serendipity['GET']['unlock']) && $serendipity['GET']['unlock'] == 'true')) {
                 serendipity_db_query("DELETE FROM {$serendipity['dbPrefix']}entryproperties WHERE (property = 'locked' OR property = 'lock_owner') AND entryid = " . (int)$id);
+                $state = 'liberate';
             } else {
                 $state = 'locked';
             }
@@ -144,18 +145,18 @@ class serendipity_event_entrycheck extends serendipity_event
         if (isset($hooks[$event])) {
 
             switch($event) {
-                case 'css_backend':
+                case 'css':
                     $eventData .= '
 
 /* entrycheck plugin start */
 
-.entrylock { margin: 1.5em auto; width: auto; text-align: center; padding: 5px; border: 1px solid yellow; }
-.entrylock a.serendipityPrettyButton { margin: 1.5em; }
+body.save_preview_body .msg_error {
+    font-size: inherit;
+}
 
 /* entrycheck plugin end */
 
 ';
-                    break;
 
                 case 'backend_entryform':
                     if (!isset($eventData['id']) || $eventData['id'] < 1) {
@@ -174,33 +175,35 @@ class serendipity_event_entrycheck extends serendipity_event
                             $locked = array('lock_owner' => $serendipity['authorid'], 'locked' => $time);
                         }
 
-                        $owner = serendipity_fetchAuthor($locked['lock_owner']);
-                        $link = '<a href="serendipity_admin.php?serendipity[action]=admin&amp;serendipity[adminModule]=entries&amp;serendipity[adminAction]=edit&amp;serendipity[id]=' . (int)$eventData['id'] . '&amp;serendipity[unlock]=true" class="serendipityPrettyButton">' . PLUGIN_EVENT_ENTRYCHECK_UNLOCK . '</a>';
-                        printf('<div class="entrylock">' . PLUGIN_EVENT_ENTRYCHECK_LOCKED . ' ' . $link . '</div>', $owner[0]['realname'], serendipity_strftime(DATE_FORMAT_SHORT, (int)$locked['locked']));
+                        if ($state != 'liberate' && !empty($locked['lock_owner'])) {
+                            $owner = serendipity_fetchAuthor($locked['lock_owner']);
+                            $link = '<a href="serendipity_admin.php?serendipity[action]=admin&amp;serendipity[adminModule]=entries&amp;serendipity[adminAction]=edit&amp;serendipity[id]=' . (int)$eventData['id'] . '&amp;serendipity[unlock]=true&amp;' . serendipity_setFormToken('url') . '" class="serendipityPrettyButton">' . PLUGIN_EVENT_ENTRYCHECK_UNLOCK . '</a>';
+                            printf('<div class="msg_info">' . PLUGIN_EVENT_ENTRYCHECK_LOCKED . ' ' . $link . '</div>', $owner[0]['realname'], serendipity_strftime(DATE_FORMAT_SHORT, (int)$locked['locked']));
+                        }
                     }
                     break;
 
                 case 'backend_entry_updertEntry':
                     if (serendipity_db_bool($this->get_config('emptyCategories', 'false') === true) && (isset($addData['categories']) && count($addData['categories']) < 1) || (isset($addData['categories'][0]) && $addData['categories'][0] == '0')) {
-                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYCATEGORIES_WARNING.'<br>';
+                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYCATEGORIES_WARNING;
                     }
 
                     if (serendipity_db_bool($this->get_config('emptyTitle', 'false') === true) && strlen($addData['title']) < 1) {
-                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYTITLE_WARNING.'<br>';
+                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYTITLE_WARNING;
                     }
 
                     if (serendipity_db_bool($this->get_config('emptyBody', 'false') === true) && strlen($addData['body']) < 1) {
-                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYBODY_WARNING.'<br>';
+                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYBODY_WARNING;
                     }
 
                     if (serendipity_db_bool($this->get_config('emptyExtended', 'false') === true) && strlen($addData['extended']) < 1) {
-                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYEXTENDED_WARNING.'<br>';
+                        $eventData[] = PLUGIN_EVENT_ENTRYCHECK_EMPTYEXTENDED_WARNING;
                     }
 
                     if ($addData['id'] > 0 && serendipity_db_bool($this->get_config('locking', 'false')) === true) {
                         $state = 'unlocked';
                         if (!$this->checkLock($state, $addData['id'])) {
-                            $eventData[] = PLUGIN_EVENT_ENTRYCHECK_LOCK_WARNING.'<br>';
+                            $eventData[] = PLUGIN_EVENT_ENTRYCHECK_LOCK_WARNING;
                         }
                     }
                     break;
